+++ lib/src/caches/simple_fifo_cache.dart
@@ -44,6 +44,12 @@ class SimpleFIFOCache<K, V> extends SimpleCache<K, V> {
     return _cache[key];
   }
 
+  /// Checks whether [key] exists in the cache.
+  ///
+  /// **This method is not thread-safe.**
+  @override
+  bool containsKey(K key) => _cache.containsKey(key);
+
   /// Stores the specified key-value pair in the cache.
   ///
   /// - If `set()` is called on an existing key, **its value is updated**.